Adjust GitHub Actions workflows to build and publish snap
* The snap's confinement mode is switched to classic. Mostly because JL
Server creates hidden and temp files in a jupyter notebook directory it
has open (e.g. for checkpoints or when it saves the file). Thus, if a
user has a notebook stored in $HOME, JLD fails to save it with a
permission denied error. It cannot be resolved with the snap's
personal-files interface as the interface allows to specify existing
paths only.